requirements.txt
alembic==1.8.1
autopep8==2.0.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-CKEditor==0.4.6
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-Testing==0.8.1
Flask-WTF==1.0.1
greenlet==2.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
Mako==1.2.3
MarkupSafe==2.1.1
Pillow==9.3.0
pycodestyle==2.9.1
PyJWT==2.6.0
python-dotenv==0.21.0
SQLAlchemy==1.4.42
tomli==2.0.1
Werkzeug==2.2.2
win32-setctime==1.1.0
WTForms==3.0.1
